+++ modules/tools/src/libifm3d_tools/discover_app.cpp
@@ -55,7 +55,8 @@ ifm3d::DiscoverApp::Run()
               std::cout << ip_address << " (Unsupported device)" << std::endl;
               continue;
             }
-          std::cout << ip_address << " (" << device_type << ")" << std::endl;
+          std::cout << ip_address << " (" << device_type << ")"
+                    << " [" << device.GetMACAddress() << "]" << std::endl;
         }
       catch (ifm3d::Error& e)
         {